Add a function for matching JSON values
This will be the lowest level matching predicate in the tool, and will
soon be accompanied by a similar method for matching JSON objects (by
key and value).

Arrays are recursed so it will be possible to check whether a JSON
object such as:

	{ "tags": [ "alpha", "beta" ] }

contains a "tags" property with the value "beta".