Add a function for matching JSON objects
This builds on the previous commit, adding a matching predicate that is
able to test whether a given JSON object matches a given key/value pair.

There is a question about whether a missing property should match an
empty string (given the requirements for this specific project). Until I
have an answer I've included a compile-time macro that can control this
behaviour.