Add minizen_db_open/minizen_db_close functions
These can be used by an application to get a "connection" to a
"database" (or in our case, stash a reference to the directory that will
contain our JSON data files).

The minizen tool is also updated to call these functions (using the
current working directory, unless overridden on the command line using a
new --data-dir argument).