Add minizen_db_get_path function
This can be used for returning a path relative to the database's data
directory. It will be used in an upcoming commit to resolve paths to
JSON data files, for which we expect to use varargs-style calls; e.g.:

	minizen_db_get_path(db, "%s.json", ...)

which saves us from having to deal with another string concatenation.

As noted in the comments this is probably an overkill; but the common
alternative:

	char path[PATH_MAX];
	snprintf(path, PATH_MAX, "%s/%s", db->dir, filename);

can leave us with a truncated path; see:
http://insanecoding.blogspot.com/2007/11/pathmax-simply-isnt.html