Cache JSON file contents
This speeds up access in case the same table needs to be queried
multiple times (which could be the case when querying related entities
in the future) at the cost of increased memory usage.

Without caching, and the ability to search for multiple key values in
one query, we could otherwise see O(ab) performance. That said, there
could be times we don't want to cache, and so ideally this would become
configurable - that's why minizen_db_get_table has been made
semantically equivalent to minizen_db_load_table.