Fix a clang warning
Clang warns about something GCC doesn't (by default):

db.c:103:32: warning: result of comparison of constant 4294967295 with expression of type 'minizen_table_t'
      (aka 'enum minizen_table') is always false [-Wtautological-constant-out-of-range-compare]
        if (table_name_to_slot(table) == -1) {

Despite the warning, our tests pass - there is one that checks this -
showing the comparison *isn't* always false.. which isn't unexpected
as the C spec says an enum is a signed int and the use of -1 is quite
standard. We just create a new enum value, with the value of -1, and
clang is happy. As are we.